LIBFCREPO-1520: Subclassing Application Controller instead of Bookmarks
There was a side effect, where subclassing the bookmarks controller
would affect the length of the table

The bookmarks controller was subclassing the catalog controller and
within the Catalog module for blacklight, there is a method that
determines what class to add to a section tag that gets loaded before
the templates.

For show methods, a class of 'col-lg-9' was added to the section tag
which would make the table take up 75% of the screen.

For all other methods, a class of 'col-md-12' was used instead, allowing
the table take up 100% of the screen.

Removed the subclassing of the bookmarks controller and added the
necessary modules to retain the functionality of the publish_jobs
controller.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1520