LIBFCREPO-1588: Add ViewComponent for listing metadata fields
With rails 7 there was an introduction to the use of ViewComponents,

Different templates can be applied to each metadata field that you want
to present within Blacklight.

So this one just parses the list that solr returns and creates a
li tag for each field.

There was a guide for using these in blacklight 7 here:
https://workshop.projectblacklight.org/v8.0.0.alpha/configuring_metadata

https://umd-dit.atlassian.net/browse/LIBFCREPO-1588