LIBFCREPO-1588: Configuration to prevent nested dicts to appear
The item__first field is a child document containing a bunch of
proxies for each page. This becomes a very nested JSON object that
the RSolr library errors on when trying to parse it.

This just configures it to be an empty string instead of a dict
because it was easier than trying to figure out how to exclude it
from the query

https://umd-dit.atlassian.net/browse/LIBFCREPO-1588