LIBFCREPO-1588: Updating "item" to "object"
In https://umd-dit.atlassian.net/browse/LIBFCREPO-1627 the top level
objects were unified under the term "object"

This was to prevent complexity in Archelon from having to deal with
multiple top level objects, such as "item", "poster", "issue" and
"letter", from being prefixed with their name on each field.

There's additonal context with the complexity it was adding before the
issue in the above ticket.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1588