LIBFCREPO-1525. Fix broken unit test prior to Rubocop fixes
Fixed a broken unit test prior to embarking on making Rubocop fixes.

This unit test is failing because the expected error message has changes
from "unexpected token" to "unexpected end of input". Since the
important part of the test is the fact that an error occurs, not the
error message itself, simplified the test by only looking for the
word "unexpected".

https://umd-dit.atlassian.net/browse/LIBFCREPO-1525