LIBFCREPO-1533. Remove UMD-custom "PingController" health check
Removed the "PingController" health check and associated "/ping"
endpoint.

This controller and endpoint are no longer needed because Rails 7.1
provides an "/up" endpoint that performs a health check, see
https://github.com/rails/rails/blob/v7.1.3.4/railties/lib/rails/health_controller.rb

Note that the Rails 7.1 HealthController *does not* do a database check.
In discussions in the August 19, 2025 DC Daily Scrum meeting, it was
decided to use the default Rails 7.1 behavior to start, and add a
database check later if it proved to be necessary.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1533