LIBFCREPO-1655. Created "display_titles" method to concat Solr titles
Moved the logic that generates the concatenated item title on the
item detail page in the "archelon_document_component.rb" into a
"display_titles" method in "solr_document.rb".

Then used the "display_titles" method to generate the titles in the
"Title" field on the table on the Publish Job detail page.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1655