LIBFCREPO-1673. Updated MimeTypes class for Solr changes
Updated the "MimeTypes" class to reflect Solr changes in the "fcrepo"
Solr core.

As the Solr response contains all relevant documents and fields,
subqueries are not used -- instead the resulting data structure is
walked to get a unique list of MIME types for the files.

The use of TerpAI for assistance in generating the test cases is hereby
acknowledged.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1673