LIBFCREPO-1673. Remove Solr/PCDM-related code from application_helper
Removed obsolete methods referencing Solr subqueries and PCMD from
the "application_helper.rb" methods.

The "*_from_subquery" methods, as well as the "value_list" method had
previously been used in the CatalogController but those usages were
removed in LIBFCREPO-1588, making the methods unnecessary.

The "unique_component_types" method was removed because there are no
uses of the method anywhere in the codebase.

The PCDM_OBJECT, PCDM_FILE, and ALLOWED_MIME_TYPE constants are no
longer referenced in the codebase, and so were removed.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1673