LIBFCREPO-1677. Modified SolrDocument for Archelon-based "Members" URLs
Modified the "members_anchor" in "solr_document.rb" to return URLS
that link to the Archelon-based pages for the "Members" of the document,
instead of to the fcrepo-based pages.

This provides consistency with the "Select Page(s) for More Options"
panel in the right sidebar, which also provides link to the pages
in Archelon, instead of in fcrepo.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1677