LIBFCREPO-1676. Added "ArchelonBacklinkComponent" view component
Created an “ArchelonBacklinkComponent” view component using the]
following files:

* app/components/archelon_backlink_component.rb
* app/components/archelon_backlink_component.erb

The “ArchelonBacklinkComponent” view component takes the fcrepo-based
URI in the configured field and queries Solr for the Solr document with
that URI (the URI is assumed to represent the “container” of the item).
The retrieved Solr document is then processed to create a “backlink” to
the container’s Archelon page:

* The fcrepo URI is used to generate a URL to the Archelon page for the
  container (using the “solr_document_url”/”solr_document_path” URL
  helpers).

* The human-readable text for the link is generated using the mechanism
  specified in the field configuration (which is typically the title of
  the container’s page).

This implementation includes two possible ways for specifying how to
generate the human-readable link test:

* “backlink_text_accessor” - Specifies the method to call on the
  “SolrDocument” class (for example, the “display_titles” method).

* “backlink_text_field” - the Solr field in the retrieved Solr document
  (for example the “page__title__txt” field).

Both of mechanisms are used in configuring the
“ArchelonBacklinkComponent” view component in the
“app/controllers/catalog_controller.rb” on the following fields:

* Page detail page - “page__member_of__uri” Solr field - Uses the
  “backlink_text_accessor” field to specify that the “display_titles”
  in the SolrDocument class be used to generate the human-readable link
  text.

* File detail page - “page__title__txt” Solr field - Uses the
  “backlink_text_field” to specify that the value of the
  “page__title__txt” in the retrieved Solr document should be used for
  the human-readable link text.

Note that each use of the “ArchelonBacklinkComponent” view component on
a particular page triggers a Solr query – in the existing
implementation, only one instance is used on each page.

The “app/components/archelon_backlink_component.erb” file contains the
ERB template for displaying the field label and link (modeled on the
“app/components/blacklight/metadata_field_component.html.erb” in the
“projectblacklight/blacklight” GitHub repository).

https://umd-dit.atlassian.net/browse/LIBFCREPO-1676