LIBFCREPO-1716. Fix "SOLR_URL" in ".env.development" config file
Changed the "SOLR_URL" environment variable in the ".env.development"
configuration file to:

```
SOLR_URL='http://fcrepo-local:8985/solr/fcrepo'
```

so that the configuration would work "out of the box" with an
Archelon running locally. This replaces a SOLR_URL setting that
was mainly intended for a Docker-based Archelon build.

Also removed a commented out SOLR_URL entry, because it is no longer
relevant.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1716