LIBFCREPO-1716. Deleted .devcontainer directory and Dockerfile.vscode
The ".devcontainer" directory was initially set up to provide a
Docker-based VS Code Dev Container environment for Archelon versions
using older Ruby versions that were not well-supported on Apple Silicon.

Now that Archelon is using Ruby 3.2, and running Archelon locally is the
preferred environment for development, removed the ".devcontainer"
directory and its contents, so that VS Code doesn't prompt to reopen the
folder in a Dev Container.

Also removed "Dockerfile.vscode" which was used as part of the
VS Code Dev Container setup, and mentions of the VS Code Dev Container
in the documentation.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1716