LIBFCREPO-1704. Enabled "Identifier" searches to use "http" or "https"
Added a "UmdSearchState" class that subclasses the default Blacklight
"SearchState" class to modify the "q" parameter in an "Identifier"
search to allow both "http" and "https" URLs to be retrieved,
irrespective of which schema is provided by the user.

Modified "app/controllers/catalog_controller.rb" to use the
"UmdSearchState" class in place of the default Blacklight "SearchState"
class.

Added test cases and documentation on Archelon customizations to the
default Blacklight search functionality.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1704