LIBFCREPO-1730. Load "application.js" as module
Modified the "base.html.erb" file to load the "application.js" file
"type=module" attribute, because the JavaScript file built by the
"esbuild" bundler tool is generated using the "format=esm" directive.

This means that the JavaScript is generated in ECMAScript Module (ESM)
format, which is the most suitable format for modern browsers.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1730